v2 with unit testing, swagger-ui and HELP.md

+++ src/main/java/com/example/onebox/service/impl/CartServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.onebox.service.impl;
 
 import com.example.onebox.controller.dto.CartItemRequest;
+import com.example.onebox.controller.dto.CartItemResponse;
 import com.example.onebox.controller.dto.CartResponse;
 import com.example.onebox.domain.Cart;
 import com.example.onebox.domain.CartItem;
@@ -10,8 +11,10 @@ import com.example.onebox.service.CartService;
 import com.example.onebox.utils.MessageUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class CartServiceImpl implements CartService {
@@ -35,7 +38,10 @@ public class CartServiceImpl implements CartService {
                 .orElseThrow(() ->
                         new IllegalArgumentException(MessageUtils.getMessage("cart.not.found")));
 
-        List<CartItem> items = cartItemRepository.findByCartId(cartId);
+        List<CartItemResponse> items = cartItemRepository.findByCartId(cartId)
+                .stream()
+                .map(this::mapCartItemResponse)
+                .collect(Collectors.toList());
 
         return CartResponse.builder()
                 .cartId(cart.getId())
@@ -46,8 +52,9 @@ public class CartServiceImpl implements CartService {
     @Override
     public CartResponse addItem(Long cartId, CartItemRequest cartItemRequest) {
 
-        if (!cartRepository.existsById(cartId))
-            throw new IllegalArgumentException(MessageUtils.getMessage("cart.not.found"));
+        Cart cart = cartRepository.findById(cartId)
+                .orElseThrow(() ->
+                        new IllegalArgumentException(MessageUtils.getMessage("cart.not.found")));
 
         CartItem cartItem = CartItem.builder()
                 .cartId(cartId)
@@ -56,13 +63,28 @@ public class CartServiceImpl implements CartService {
                 .build();
 
         cartItemRepository.save(cartItem);
+        cartRepository.save(cart); // update modification date
 
         return getCart(cartId);
     }
 
     @Override
+    @Transactional
     public void deleteCart(Long cartId) {
+
+        if (!cartRepository.existsById(cartId))
+            throw new IllegalArgumentException(MessageUtils.getMessage("cart.not.found"));
+
         cartItemRepository.deleteAllByCartId(cartId);
         cartRepository.deleteById(cartId);
+    }
+
+    private CartItemResponse mapCartItemResponse(CartItem cartItem) {
+        return CartItemResponse.builder()
+                .amount(cartItem.getAmount())
+                .description(cartItem.getDescription())
+                .cartItemId(cartItem.getId())
+                .cartId(cartItem.getCartId())
+                .build();
     }
 }
